Finish implementing smooth scroll using rAF.
If a browser does not support rAF, it will fallback to a
non-animated version and just set scrollTop to the correct
position. While the fallback may be used more on older mobile
devices, it will be more efficient in general, because the mobile
devices will decide the appropriate frame rate of the animation.