Ran through http://www.javascriptlint.com/online_lint.php
Some of these changes are probably debatable. Line length vs... I forget what it said.
I tried Google Closure Linter, but it didn't change it except for changing " to ' and @returns JSDoc to @return so I didn't include the changes it made
(mainly because I accidently deleted it after using a diff tool)